+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 CHANGELOG
 =========
 
-5.4, 2022-09-15
+5.4, 2022-09-16
 ---------------
+- add database account as one entry among contacts in contacts-tab
 - add configuration flag to cache shared media on disk (issue #112)
 - skip trying to download old shared content guaranteed to be unavailable